Enable chain client certs
When using a intermediate cert, the intermediate CA should be appended
to the client cert to allow the server to verify the signature.

Before we only used the first certificate, so things were not working
correctly.

This patched changes a bit the nginx module, plus added a few test
scripts.

Integration test to be made on APICast QE testsuite.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>