Adding a helper function to check if should be verified.
When handshake happens, we also need to check the CN. Something that
it's not done because u->conf->verify maybe is 0. With this change, a
new function is exported to handle this case on nginx-upstream.

Lines that need to be checked:
https://github.com/nginx/nginx/blob/02bd43d05b6f7803597d8453d9848b767dc4a323/src/http/ngx_http_upstream.c#L1786-L1789

Changes that need to be done on Openresty:
```

static void
ngx_http_upstream_ssl_handshake(ngx_http_request_t *r, ngx_http_upstream_t *u,
    ngx_connection_t *c)
{
    long  rc;

    if (c->ssl->handshaked) {

        if (u->conf->ssl_verify || ngx_http_upstream_should_verified(r) == NGX_OK) {
```

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>